+++ packages/web3/CHANGELOG.md
@@ -1,5 +1,23 @@
 # pelican-web3-lib
 
+## 1.26.8
+
+### Patch Changes
+
+- Updated dependencies
+  - pelican-web3-lib-assets@1.0.8
+  - pelican-web3-lib-common@1.0.8
+  - pelican-web3-lib-icons@1.0.8
+
+## 1.26.7
+
+### Patch Changes
+
+- Updated dependencies
+  - pelican-web3-lib-assets@1.0.7
+  - pelican-web3-lib-common@1.0.7
+  - pelican-web3-lib-icons@1.0.7
+
 ## 1.26.6
 
 ### Patch Changes